Clustermap audit (#1732)
